feat(product): Adds on-sale filtering and multi-category support
Introduces a new `onSale` query parameter to the `getProducts` endpoint. When `onSale=true`, products with a discount greater than 0 or a 'sale' badge type will be returned.

Enhances the `category` and `subcategory` filters to support multiple values. Users can now pass an array of categories or subcategories, which will utilize MongoDB's `$in` operator for broader filtering capabilities.